feat(translate_xml_docs): add script to translate from within xml files with <doc> tags
This script uses the mux/demux threading pattern to read the text from within <doc> tags, split it on newlines, feed it to the running subprocess (e.g. fairseq), read it bak line-by-line, reinsert it into the <doc> tags, and write the entire document to the output file. For now, I have used ET.parse, which I believe reads the WHOLE XML FILE INTO MEMORY. This is because I know we currently only have small files to translate.